requirements.txt
fastapi[standard]
uvicorn[standard]
python-multipart
python-dotenv
python-docx
pandas
openpyxl
Pillow
pytesseract
pymupdf
python-dotenv
pinecone
langchain
openai
